added tests and a few modifications to the screens

+++ app/src/main/java/com/github/wanderwise_inc/app/ui/list_itineraries/OverviewScreen.kt
@@ -81,8 +81,10 @@ fun DisplayOverviewItineraries(
                       itinerary.description?.contains(searchQuery, ignoreCase = true) ?: false
                 }
                 .filter { itinerary ->
-                  val price = itinerary.price.toFloat()
-                  price in sliderPositionPriceState.value.start..sliderPositionPriceState.value.endInclusive
+                  val price = itinerary.price
+                  price in
+                      sliderPositionPriceState.value.start..sliderPositionPriceState.value
+                              .endInclusive
                 }
                 .filter { itinerary ->
                   val time = itinerary.time.toFloat()
